+++ src/ModelManager.cpp
@@ -1,7 +1,8 @@
 #include "ModelManager.h"
 
-#include "Colors.h"
 #include <imgui.h>
+
+#include "Colors.h"
 
 int ModelManager::sModelId = 0;
 
@@ -9,8 +10,9 @@ void ModelManager::AddModel(std::shared_ptr<Model> model) {
   mModels.push_back(model);
   std::string name = model->GetPath() + "_" + std::to_string(sModelId++);
   mModelNames.push_back(name);
-  mModelSegmentedColors = Colors::GenerateSegmentedColors(mModels.size());
-  mSelectedModelId = mModels.size() - 1;
+  mModelSegmentedColors =
+      Colors::GenerateSegmentedColors(static_cast<int>(mModels.size()));
+  mSelectedModelId = static_cast<int>(mModels.size()) - 1;
   Logger::getInstance().Info("ModelManager: Added model " + model->GetPath());
 }
 
@@ -22,7 +24,8 @@ void ModelManager::Remove() {
   mModelNames.erase(mModelNames.begin() + mSelectedModelId);
   if (mModels.size() > 0) {
     mSelectedModelId = 0;
-    mModelSegmentedColors = Colors::GenerateSegmentedColors(mModels.size());
+    mModelSegmentedColors =
+        Colors::GenerateSegmentedColors(static_cast<int>(mModels.size()));
   } else {
     mSelectedModelId = -1;
   }
@@ -38,7 +41,7 @@ void ModelManager::ShowModels() {
       const bool is_selected = (mSelectedModelId == n);
       if (ImGui::Selectable(mModelNames[n].c_str(), is_selected) &&
           mSelectedModelId != n) {
-        mSelectedModelId = n;
+        mSelectedModelId = static_cast<int>(n);
       }
       if (is_selected)
         ImGui::SetItemDefaultFocus();
